pyproject.toml
[project]
name = "teste-uv"
version = "0.1.0"
description = "teste de script uv"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.120.1",
    "mysql>=0.0.3",
    "mysql-connector>=2.2.9",
    "pandas>=2.3.3",
    "pydantic-settings>=2.11.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
    "taskipy>=1.14.1",
]

[tool.taskipy.tasks]
s = "fastapi dev main.py"